+++ changelog.md
@@ -5,6 +5,7 @@ Codion Change Log
 ### is.codion.common.core
 - EventObserver.addDataListener() renamed addConsumer(), related renaming.
 - Value.Builder added, replaces a couple of static factory methods.
+- Value.Builder.link() added.
 ### is.codion.framework.domain
 - QueryKeyGenerator, SequenceKeyGenerator, now overwrite a current value instead of silently skipping it.
 - AbstractQueriedKeyGenerator.selectAndPopulate() primary key column definition parameter removed.